Implementering af lists4
