+++ .idea/misc.xml
@@ -3,8 +3,22 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/Medarbejder Support2/AndroidStudioProjects/Afleverings-Android-projekt/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.1165" />
+        <entry key="..\:/Users/Medarbejder Support2/AndroidStudioProjects/Afleverings-Android-projekt/app/src/main/res/drawable/border.xml" value="0.1165" />
+        <entry key="..\:/Users/Medarbejder Support2/AndroidStudioProjects/Afleverings-Android-projekt/app/src/main/res/drawable/ic_dashboard_black_24dp.xml" value="0.1165" />
+        <entry key="..\:/Users/Medarbejder Support2/AndroidStudioProjects/Afleverings-Android-projekt/app/src/main/res/drawable/ic_home_black_24dp.xml" value="0.1165" />
+        <entry key="..\:/Users/Medarbejder Support2/AndroidStudioProjects/Afleverings-Android-projekt/app/src/main/res/drawable/ic_launcher_background.xml" value="0.1165" />
+        <entry key="..\:/Users/Medarbejder Support2/AndroidStudioProjects/Afleverings-Android-projekt/app/src/main/res/drawable/ic_notifications_black_24dp.xml" value="0.1165" />
+        <entry key="..\:/Users/Medarbejder Support2/AndroidStudioProjects/Afleverings-Android-projekt/app/src/main/res/layout/activity_add_product.xml" value="0.33" />
+        <entry key="..\:/Users/Medarbejder Support2/AndroidStudioProjects/Afleverings-Android-projekt/app/src/main/res/layout/activity_main.xml" value="0.14659367396593673" />
         <entry key="..\:/Users/Medarbejder Support2/AndroidStudioProjects/Afleverings-Android-projekt/app/src/main/res/layout/activity_menu_view.xml" value="0.10885416666666667" />
         <entry key="..\:/Users/Medarbejder Support2/AndroidStudioProjects/Afleverings-Android-projekt/app/src/main/res/layout/activity_my_shoping_list.xml" value="0.25" />
+        <entry key="..\:/Users/Medarbejder Support2/AndroidStudioProjects/Afleverings-Android-projekt/app/src/main/res/layout/activity_pantry.xml" value="0.12552083333333333" />
+        <entry key="..\:/Users/Medarbejder Support2/AndroidStudioProjects/Afleverings-Android-projekt/app/src/main/res/layout/activity_products_add.xml" value="0.12552083333333333" />
+        <entry key="..\:/Users/Medarbejder Support2/AndroidStudioProjects/Afleverings-Android-projekt/app/src/main/res/layout/fragment_pantry.xml" value="0.23958333333333334" />
+        <entry key="..\:/Users/Medarbejder Support2/AndroidStudioProjects/Afleverings-Android-projekt/app/src/main/res/layout/products.xml" value="0.33" />
+        <entry key="..\:/Users/Medarbejder Support2/AndroidStudioProjects/Afleverings-Android-projekt/app/src/main/res/layout/recyclerview.xml" value="0.25" />
+        <entry key="..\:/Users/Medarbejder Support2/AndroidStudioProjects/Afleverings-Android-projekt/app/src/main/res/menu/bottom_nav_menu.xml" value="0.25" />
         <entry key="..\:/Users/Troel/AndroidStudioProjects/MyPantry/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.1605" />
         <entry key="..\:/Users/Troel/AndroidStudioProjects/MyPantry/app/src/main/res/drawable/ic_dashboard_black_24dp.xml" value="0.1605" />
         <entry key="..\:/Users/Troel/AndroidStudioProjects/MyPantry/app/src/main/res/drawable/ic_home_black_24dp.xml" value="0.1605" />
