+++ app/build.gradle
@@ -46,6 +46,8 @@ dependencies {
     implementation 'androidx.navigation:navigation-ui:2.4.1'
     implementation'androidx.recyclerview:recyclerview:1.2.1'
     implementation 'androidx.room:room-common:2.4.2'
+    implementation 'androidx.leanback:leanback:1.0.0'
+    implementation 'com.github.bumptech.glide:glide:4.11.0'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
